+++ src/controllers/component.controller.js
@@ -2,13 +2,8 @@
 require('dotenv').config();
 //Services
 const {
-  getComponentByIdService,
-  getAllComponentLikeCodeService,
-  getAllComponentLikeImageService,
   getAllComponentLikePartNumberService,
   getAllComponentLikeCategoryAndMakerService,
-  getAllComponentWithBipolarTransistorService,
-  getAllComponentWithAllModelsService,
   getAllComponentLikeDescriptionService,
   getAllComponentLikeStockMaxService,
   getAllComponentLikeStockService,
@@ -45,280 +40,6 @@ const ORDER_AT_DESCRIPTION_VALUE_ERROR =
   paginationDescriptionValueError.ORDER_AT_DESCRIPTION_VALUE_ERROR;
 let msgResponse;
 let msgLog;
-
-/**
- * @description get all paginated components list with all models/table from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllWithAllModelsComponentController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-
-    componentList = await getAllComponentWithAllModelsService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case ORDER_BY_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_BY_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case ORDER_AT_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_AT_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error:
-            'Bad request, failed to get all paginated components list and bipolar_transistor according to all attributes.',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList[0]?.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res
-            .status(OK_CODE)
-            .send({ ok: 'No items found according to the all models.' });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse = 'ERROR in getAllWithAllModelsComponentController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
-
-/**
- * @description get a component based its id from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getComponentByIdController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-
-    componentList = await getComponentByIdService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error:
-            'Bad request, could not get a component with the requested id. Check if the component exist into the database',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res
-            .status(OK_CODE)
-            .send({ ok: 'No items found according to the id.' });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse = 'ERROR in getComponentByIdController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
-
-/**
- * @description get all paginated components according to code from database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentLikeCodeController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-
-    componentList = await getAllComponentLikeCodeService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case ORDER_BY_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_BY_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case ORDER_AT_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_AT_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error:
-            'Bad request, could not get all paginated list components according the code.',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList[0]?.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res
-            .status(OK_CODE)
-            .send({ ok: 'No items found according to the code.' });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse = 'ERROR in getAllComponentLikeCodeController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
-
-/**
- * @description get all paginated components according to image from database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentLikeImageController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-
-    componentList = await getAllComponentLikeImageService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case ORDER_BY_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_BY_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case ORDER_AT_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_AT_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error:
-            'Bad request, could not get all paginated list components according the imagen.',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList[0]?.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res
-            .status(OK_CODE)
-            .send({ ok: 'No items found according to the image.' });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse = 'ERROR in getAllComponentLikeImageController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
 
 /**
  * @description get all paginated components according to part number from database
@@ -971,10 +692,6 @@ const getAllComponentLikePriceMinMaxController = async (req, res) => {
 };
 
 module.exports = {
-  getAllWithAllModelsComponentController,
-  getComponentByIdController,
-  getAllComponentLikeCodeController,
-  getAllComponentLikeImageController,
   getAllComponentLikePartNumberController,
   getAllComponentLikeCategoryAndMakerController,
   getAllComponentLikeDescriptionController,
