+++ src/services/component.service.js
@@ -2,12 +2,6 @@
 const { Op } = require('sequelize');
 //Models
 const { Component } = require('../models/sequelize/component');
-const { ComponentDetail } = require('../models/sequelize/component-detail');
-const { BipolarTransistor } = require('../models/sequelize/bipolar-transistor');
-const { MosfetTransistor } = require('../models/sequelize/mosfet-transistor');
-const {
-  ElectrolycticCapacitor,
-} = require('../models/sequelize/electrolytic_capacitor');
 //Enums
 const { statusName } = require('../enums/database/status');
 const { checkErrors } = require('../helpers/sequelize/errors');
@@ -24,13 +18,9 @@ const ORDER_AT_NAME_VALUE_ERROR =
   paginationNameValueError.ORDER_AT_NAME_VALUE_ERROR;
 //componentss
 let componentList;
-let component;
 //params
 let queryStrParams;
-let idParam;
-let codigoParam;
 let params;
-let imagenParam;
 let nroPiezaParam;
 let categoriaParam;
 let descripcionParam;
@@ -48,323 +38,6 @@ let orderBy;
 let orderAt;
 let order;
 let msg;
-
-/**
- * @description get all paginated components with all models from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentWithAllModelsService = async (req, res) => {
-  try {
-    componentList = null;
-    queryStrParams = null;
-    //Pagination
-    pageSizeNro = 5;
-    pageNro = 0;
-    orderBy = 'id';
-    orderAt = 'ASC';
-    msg = null;
-
-    //-- start with pagination  ---
-    queryStrParams = req.query;
-
-    if (queryStrParams != (null && undefined)) {
-      //Pagination
-      pageSizeNro = queryStrParams.limit
-        ? parseInt(queryStrParams.limit)
-        : pageSizeNro;
-      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
-      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
-      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
-    }
-
-    orderBy = await checkOrderBy(orderBy);
-
-    if (orderBy == (null || undefined)) {
-      return ORDER_BY_NAME_VALUE_ERROR;
-    }
-
-    orderAt = await checkOrderAt(orderAt);
-
-    if (orderAt == (undefined || null)) {
-      return ORDER_AT_NAME_VALUE_ERROR;
-    }
-
-    order = [[orderBy, orderAt]];
-
-    //-- end with pagination  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findAll({
-        attributes: {},
-        where: {},
-        include: [
-          { model: ComponentDetail, required: false },
-          { model: BipolarTransistor, required: false },
-          { model: MosfetTransistor, required: false },
-          { model: ElectrolycticCapacitor, required: false },
-        ],
-        limit: pageSizeNro,
-        offset: pageNro,
-        order: order,
-        raw: true,
-        nest: true,
-      })
-        .then(async (componentItems) => {
-          componentList = componentItems;
-        })
-        .catch(async (error) => {
-          msg = `Error in getAllComponentWithAllModelsService() function when trying to get all paginated components. Caused by ${error}`;
-          console.log(msg);
-
-          componentList = await checkErrors(error, error.name);
-        });
-    } else {
-      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getAllComponentWithAllModelsService() function. Caused by ${error}`;
-    console.log(msg);
-
-    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return componentList;
-};
-
-/**
- * @description get a component according to its identifier from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getComponentByIdService = async (req, res) => {
-  try {
-    component = null;
-    params = null;
-    idParam = null;
-    msg = null;
-    //Pagination
-    msg = null;
-
-    //-- start with params ---
-    params = req.params;
-
-    if (params != (null && undefined)) {
-      idParam = params.id ? parseInt(params.id) : null;
-    }
-    //-- end with params  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findByPk(idParam, {
-        attributes: {},
-        raw: true,
-        nest: true,
-      })
-        .then(async (componentItem) => {
-          component = componentItem;
-        })
-        .catch(async (error) => {
-          msg = `Error in getComponentByIdService() function when trying to get a component by id. Caused by ${error}`;
-          console.log(msg);
-
-          component = await checkErrors(error, error.name);
-        });
-    } else {
-      component = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getComponentByIdService() function. Caused by ${error}`;
-    console.log(msg);
-
-    component = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return component;
-};
-
-/**
- * @description get all paginated components list according to its code from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentLikeCodeService = async (req, res) => {
-  try {
-    componentList = null;
-    codigoParam = null;
-    queryStrParams = null;
-    //Pagination
-    pageSizeNro = 5;
-    pageNro = 0;
-    orderBy = 'id';
-    orderAt = 'ASC';
-    msg = null;
-
-    //-- start with params ---
-    params = req.params;
-
-    if (params != (null && undefined)) {
-      codigoParam = params.codigo ? params.codigo : codigoParam;
-    }
-    //-- end with params  ---
-
-    //-- start with pagination  ---
-    queryStrParams = req.query;
-
-    if (queryStrParams != (null && undefined)) {
-      //Pagination
-      pageSizeNro = queryStrParams.limit
-        ? parseInt(queryStrParams.limit)
-        : pageSizeNro;
-      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
-      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
-      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
-    }
-
-    orderBy = await checkOrderBy(orderBy);
-
-    if (orderBy == (null || undefined)) {
-      return ORDER_BY_NAME_VALUE_ERROR;
-    }
-
-    orderAt = await checkOrderAt(orderAt);
-
-    if (orderAt == (undefined || null)) {
-      return ORDER_AT_NAME_VALUE_ERROR;
-    }
-
-    order = [[orderBy, orderAt]];
-
-    //-- end with pagination  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findAll({
-        attributes: {},
-        where: {
-          codigo: {
-            [Op.like]: `%${codigoParam}%`,
-          },
-        },
-        limit: pageSizeNro,
-        offset: pageNro,
-        order: order,
-        raw: true,
-        nest: true,
-      })
-        .then(async (componentItems) => {
-          componentList = componentItems;
-        })
-        .catch(async (error) => {
-          msg = `Error in getComponentLikeCodigoService() function when trying to get a component by codigo. Caused by ${error}`;
-          console.log(msg);
-
-          componentList = await checkErrors(error, error.name);
-        });
-    } else {
-      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getComponentLikeCodigoService() function. Caused by ${error}`;
-    console.log(msg);
-
-    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return componentList;
-};
-
-/**
- * @description get all paginated components list according to its imagen from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentLikeImageService = async (req, res) => {
-  try {
-    componentList = null;
-    imagenParam = null;
-    queryStrParams = null;
-    //Pagination
-    pageSizeNro = 5;
-    pageNro = 0;
-    orderBy = 'id';
-    orderAt = 'ASC';
-    msg = null;
-
-    //-- start with params ---
-    params = req.params;
-
-    if (params != (null && undefined)) {
-      imagenParam = params.imagen ? params.imagen : imagenParam;
-    }
-    //-- end with params  ---
-
-    //-- start with pagination  ---
-    queryStrParams = req.query;
-
-    if (queryStrParams != (null && undefined)) {
-      //Pagination
-      pageSizeNro = queryStrParams.limit
-        ? parseInt(queryStrParams.limit)
-        : pageSizeNro;
-      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
-      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
-      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
-    }
-
-    orderBy = await checkOrderBy(orderBy);
-
-    if (orderBy == (null || undefined)) {
-      return ORDER_BY_NAME_VALUE_ERROR;
-    }
-
-    orderAt = await checkOrderAt(orderAt);
-
-    if (orderAt == (null || undefined)) {
-      return ORDER_AT_NAME_VALUE_ERROR;
-    }
-
-    order = [[orderBy, orderAt]];
-
-    //-- end with pagination  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findAll({
-        attributes: {},
-        where: {
-          imagen: {
-            [Op.like]: `%${imagenParam}%`,
-          },
-        },
-        limit: pageSizeNro,
-        offset: pageNro,
-        order: order,
-        raw: true,
-        nest: true,
-      })
-        .then(async (componentItems) => {
-          componentList = componentItems;
-        })
-        .catch(async (error) => {
-          msg = `Error in getAllComponentLikeImageService() function when trying to get a component by imagen. Caused by ${error}`;
-          console.log(msg);
-
-          componentList = await checkErrors(error, error.name);
-        });
-    } else {
-      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getAllComponentLikeImageService() function. Caused by ${error}`;
-    console.log(msg);
-
-    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return componentList;
-};
 
 /**
  * @description get all paginated components list according to its part number from the database
@@ -1203,10 +876,6 @@ const getAllComponentLikePriceMinMaxService = async (req, res) => {
 };
 
 module.exports = {
-  getAllComponentWithAllModelsService,
-  getComponentByIdService,
-  getAllComponentLikeCodeService,
-  getAllComponentLikeImageService,
   getAllComponentLikePartNumberService,
   getAllComponentLikeCategoryAndMakerService,
   getAllComponentLikeDescriptionService,
