+++ src/services/component/get-by-id.js
@@ -1,0 +1,67 @@
+//Models
+const { Component } = require('../../models/sequelize/component');
+//Enums
+const { statusName } = require('../../enums/database/status');
+const { checkErrors } = require('../../helpers/sequelize/errors');
+//components
+let component;
+//params
+let idParam;
+let params;
+//pagination
+let msg;
+
+/**
+ * @description get a component according to its identifier from the database
+ * @param {any} req any type
+ * @param {any} res any type
+ * @returns a json object with the transaction performed
+ * @example
+ */
+const getComponentByIdService = async (req, res) => {
+  try {
+    component = null;
+    params = null;
+    idParam = null;
+    msg = null;
+    //Pagination
+    msg = null;
+
+    //-- start with params ---
+    params = req.params;
+
+    if (params != (null && undefined)) {
+      idParam = params.id ? parseInt(params.id) : null;
+    }
+    //-- end with params  ---
+
+    if (Component != (null && undefined)) {
+      await Component.findByPk(idParam, {
+        attributes: {},
+        raw: true,
+        nest: true,
+      })
+        .then(async (componentItem) => {
+          component = componentItem;
+        })
+        .catch(async (error) => {
+          msg = `Error in getComponentByIdService() function when trying to get a component by id. Caused by ${error}`;
+          console.log(msg);
+
+          component = await checkErrors(error, error.name);
+        });
+    } else {
+      component = await checkErrors(null, statusName.CONNECTION_REFUSED);
+    }
+  } catch (error) {
+    msg = `Error in getComponentByIdService() function. Caused by ${error}`;
+    console.log(msg);
+
+    component = await checkErrors(error, statusName.CONNECTION_ERROR);
+  }
+  return component;
+};
+
+module.exports = {
+  getComponentByIdService,
+};
